Make loadout use cookies
